#8 working example in README and package renaming
