+++ tests/jubjub_test.py
@@ -2,8 +2,8 @@ import unittest
 
 from os import urandom
 
-from zokrates_pycrypto.fields import BLS12_381Field as FQ
-from zokrates_pycrypto.curves import JubJub
+from znakes.fields import BLS12_381Field as FQ
+from znakes.curves import JubJub
 
 
 JUBJUB_C = JubJub.JUBJUB_C
